+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- test\Start-þnameþRunspace.ps1  
+ Helper for starting a new runspace with the current development state
+- Working example code in basic functions
+### Changed
+- Search'n'Replace of þnameþ against long versions
 
 ## [1.0.3] - 2023-03-06
 ### Fixed